Windows porting begin
Add MSVC CMake presets and split the build logic into smaller CMake modules.
Configure Boost and OpenSSL through explicit CMake roots.
Add Windows implementations for filesystem, sockets, timers, and threading.
Use native Windows threads, mutexes, condition variables, and affinity.
Preserve Linux-specific filesystem and threading behavior.
Fix MSVC export/import issues for public framework types.
Normalize SIZE and LSIZE usage across file and buffer APIs.
Update tests for the adjusted API and Windows behavior.

+++ src/time/datetime.h
@@ -36,7 +36,7 @@ namespace crb
         DateTime();
 
         // Construct a DateTime object from POSIX timespec data
-        DateTime(uint32_t seconds, uint32_t nanoseconds);
+        DateTime(int64_t seconds, int64_t nanoseconds);
 
         // Check if the instance is valid
         bool isValid() const;
@@ -116,7 +116,7 @@ namespace crb
 
         // Set the time according to the seconds and nanoseconds (timespec struct).
         // Please note that nanosecond precision is lost
-        DateTime& fromTimespec(uint32_t seconds, uint32_t nanoseconds = 0);
+        DateTime& fromTimespec(int64_t seconds, int64_t nanoseconds = 0);
 
         // Return the current system time
         static DateTime current();
